Move zeroReport util to js test folder to align with existing js tests and update imports across app.
