+++ tests/js/utils/zeroReports.test.js
@@ -20,7 +20,7 @@
  * Internal dependencies
  */
 import { replaceValuesOrRemoveRowForDateRangeInAnalyticsReport } from './zeroReports';
-import mockedReportResponse from '../../assets/js/modules/analytics-4/utils/__fixtures__/mocked-report-multiple-distinct-date-ranges.json';
+import mockedReportResponse from './../../../assets/js/modules/analytics-4/utils/__fixtures__/mocked-report-multiple-distinct-date-ranges.json';
 
 describe( 'zeroReports', () => {
 	describe( 'replaceValuesOrRemoveRowForDateRangeInAnalyticsReport', () => {
